fixup! denylist: add rhel10 for relevant entries
There were two mpath iso install tests that were dropped here
that need to be added back.

Also change the link for the tracker for the team and ifcfg denylist
entries because the entries can be dropped when the submodule bump
happens. The other tracker issue is irrelevant because we changed
the definition of the tests.